bootLongMSEPsi

Compute MSE in computing \(\psi\) = two-sided probability with different block sizes.

bootLongMSEPsi(ps, main_factor, time_var, subjectID_var, sampleID_var, b,
  R, RR, qj, Wj, Khat.obs = NULL, T.obs.full = NULL, ncores,
  compStatParallel = FALSE)

Arguments

ps
phyloseq object.

main_factor
Character string. The name of the covariate variable.

time_var
Character string. The name of the time variable.

subjectID_var
Character string. The name of the subject ID variable.

sampleID_var
Character string. The name of the sample ID variable.

b
A numeric. The block size to account for the dependence within-subject.

R
A numeric. The number of block bootstrap realizations.

RR
A numeric. The number of double block bootstrap realizations.

qj
A numeric vector. The number of repeated observations for j-th subject.

Wj
A numeric vector. The number of repeated observations in the subset for j-th subject

Khat.obs
A numeric vector. The second element of the output of bootLongPsi evaluated using the initial block size and full data.

T.obs.full
A numeric vector. The numeric vector of observed statistic.

ncores
A numeric. The number of cores to use in mclapply.

compStatParallel
A logical. True is used to use parallel in computeStat and disable parallel in bootLongPsi or bootLongMethod.

Value

A list of MSE computed with given block size b, Khat with all subsamples, Khat with initial block size.